# backend/docs/sales.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FashionStore — Ventas y checkout",
    "version": "1.0.0",
    "description": "Ventas presenciales y pedidos WEB/MOBILE. JWT obligatorio. El checkout retiene stock durante CHECKOUT_HOLD_MINUTES (15 por defecto); Stripe se integra mediante el módulo de Pagos. No existe endpoint público de aprobación electrónica."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/sales/{id}": {
      "get": {
        "operationId": "getSale",
        "summary": "Consultar venta propia o de sucursal autorizada",
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación correcta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Sale"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente o inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permisos insuficientes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente o carrito ajeno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de stock, cotización, estado, idempotencia o referencia de pago",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SaleLine": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "sizeId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "colorId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "quantity": {
            "type": "integer"
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "sizeName": {
            "type": "string",
            "nullable": true
          },
          "colorName": {
            "type": "string",
            "nullable": true
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          },
          "discount": {
            "type": "number",
            "minimum": 0,
            "description": "Descuento por unidad."
          },
          "netUnitPrice": {
            "type": "number",
            "minimum": 0
          },
          "subtotal": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "productId",
          "sizeId",
          "colorId",
          "quantity",
          "productName",
          "sizeName",
          "colorName",
          "unitPrice",
          "discount",
          "netUnitPrice",
          "subtotal"
        ]
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "saleId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "method": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD",
              "QR",
              "BANK_TRANSFER",
              "GATEWAY"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "IN_STORE",
              "ELECTRONIC"
            ]
          },
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "VOIDED",
              "REFUNDED"
            ]
          },
          "externalReference": {
            "type": "string",
            "nullable": true
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha del registro; solo representa cobro confirmado cuando status=APPROVED."
          }
        },
        "required": [
          "id",
          "saleId",
          "method",
          "type",
          "amount",
          "status",
          "externalReference",
          "paidAt"
        ]
      },
      "Sale": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "clientId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647,
            "nullable": true
          },
          "employeeId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647,
            "nullable": true
          },
          "branchId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647,
            "nullable": true
          },
          "reservationId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647,
            "nullable": true
          },
          "cartId": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647,
            "nullable": true
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "stockReserved": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "soldAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "channel": {
            "type": "string",
            "enum": [
              "IN_STORE",
              "WEB",
              "MOBILE"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_PAYMENT",
              "PAID",
              "COMPLETED",
              "CANCELLED",
              "REFUNDED"
            ]
          },
          "total": {
            "type": "number",
            "minimum": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleLine"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Payment"
            }
          },
          "branch": {
            "type": "object",
            "nullable": true,
            "description": "id, name, city, address de la sucursal."
          },
          "client": {
            "type": "object",
            "nullable": true,
            "description": "id, userId y user con name/email."
          },
          "employee": {
            "type": "object",
            "nullable": true,
            "description": "id y user con name."
          }
        },
        "required": [
          "id",
          "clientId",
          "employeeId",
          "branchId",
          "reservationId",
          "cartId",
          "currency",
          "stockReserved",
          "expiresAt",
          "confirmedAt",
          "cancellationReason",
          "soldAt",
          "updatedAt",
          "channel",
          "status",
          "total",
          "items",
          "payments",
          "branch",
          "client",
          "employee"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "statusCode": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "path": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      }
    }
  }
}